Delete the read bound, and everything that existed to serve it
Zero callers. Every consumer of a thread read - model history refresh, dispatch
context, thread summaries, export - treats what it receives as the whole thread
and cannot tell a truncated read from a short one, so the budget was never an
abstraction anyone could use correctly. Keeping it would have shipped dead
weight plus a schema column, a migration, and the runtime cost of maintaining
both.

The win was always the collapse, which loses nothing: one surviving edit per
message instead of every edit ever sent, 2,021 rows to 41 on a real thread.
Truncation only ever removed the oldest half of a caller's context.

Removing it takes the two-phase selection with it, because without pricing
there is nothing to select on. The read is one query again, and the CTE no
longer needs to distinguish a present original from an absent one, only to rank
and to compare senders. Also gone: event_bytes with its column, migration and
backfill; the sidecar byte accounting added two commits ago; ThreadReadBudget,
ThreadWindowRead, get_thread_window; the truncation diagnostics.

Kept: the sender column, every edit-integrity fix, orphaned-edit synthesis, the
stale-export refusal, and is_full_history for the sidecar-hydration case it
actually describes. The write-time pruning contract the deleted module carried
now lives next to the query it would replace.

benchmark_thread_history_reuse.py goes too: it imports thread_resolution_reuse,
which this PR deleted several commits ago, so it has been unrunnable since.

src: +555 -939.
