+++ src/mindroom/matrix/cache/event_cache.py
@@ -5,8 +5,6 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, Protocol, runtime_checkable
 
-from .thread_read_window import UNBOUNDED_THREAD_READ, ThreadReadBudget, ThreadWindowRead
-
 if TYPE_CHECKING:
     from collections.abc import Collection
 
@@ -82,27 +80,11 @@ class ConversationEventCache(Protocol):
         self,
         room_id: str,
         thread_id: str,
-        *,
-        budget: ThreadReadBudget = UNBOUNDED_THREAD_READ,
     ) -> list[dict[str, Any]] | None:
-        """Return cached events for one thread sorted by timestamp.
+        """Return every message in one thread, sorted by timestamp.
 
-        A bounded budget returns the newest messages that fit, each with its latest edit, plus the
-        thread root. Selection never reads a payload it does not return.
-        """
-
-    async def get_thread_window(
-        self,
-        room_id: str,
-        thread_id: str,
-        *,
-        budget: ThreadReadBudget = UNBOUNDED_THREAD_READ,
-    ) -> ThreadWindowRead:
-        """Return one thread's window plus whether the budget left messages out.
-
-        Callers that label history as complete must use this rather than ``get_thread_events``:
-        a truncated window presented as full history makes completeness-dependent planning run on
-        a partial thread.
+        Collapsed, not truncated: each message arrives with the single edit that currently wins
+        it rather than with every edit it ever received. No message is left out.
         """
 
     async def get_recent_room_thread_ids(self, room_id: str, *, limit: int) -> list[str]:
