+++ src/mindroom/matrix/cache/postgres_event_cache.py
@@ -29,7 +29,6 @@ from .thread_cache_state import (
     incoming_thread_invalidation_takes_precedence,
     replacement_validated_at,
 )
-from .thread_read_window import UNBOUNDED_THREAD_READ, ThreadReadBudget, ThreadWindowRead
 
 if TYPE_CHECKING:
     from collections.abc import AsyncIterator, Awaitable, Callable, Collection
@@ -293,7 +292,6 @@ async def _create_postgres_event_cache_schema(db: AsyncConnection) -> None:
             room_id TEXT NOT NULL,
             origin_server_ts BIGINT NOT NULL,
             event_json TEXT NOT NULL,
-            event_bytes BIGINT NOT NULL DEFAULT 0,
             sender TEXT NOT NULL DEFAULT '',
             cached_at DOUBLE PRECISION NOT NULL,
             write_seq BIGINT NOT NULL DEFAULT nextval('mindroom_event_cache_write_seq'),
@@ -302,14 +300,8 @@ async def _create_postgres_event_cache_schema(db: AsyncConnection) -> None:
         """,
     )
     # Pre-existing deployments keep their table; the column is added in place and the version-4
-    # migration backfills it, so no row is left at the default size a bounded read would treat
-    # as free.
-    await db.execute(
-        """
-        ALTER TABLE mindroom_event_cache_events
-        ADD COLUMN IF NOT EXISTS event_bytes BIGINT NOT NULL DEFAULT 0
-        """,
-    )
+    # migration backfills it per namespace, so no row is left at the '' default that would make
+    # every event look like it came from the same account.
     await db.execute(
         """
         ALTER TABLE mindroom_event_cache_events
@@ -1270,30 +1262,17 @@ class PostgresEventCache:
         self,
         room_id: str,
         thread_id: str,
-        *,
-        budget: ThreadReadBudget = UNBOUNDED_THREAD_READ,
     ) -> list[dict[str, Any]] | None:
         """Return cached events for one thread sorted by timestamp."""
-        return (await self.get_thread_window(room_id, thread_id, budget=budget)).events
-
-    async def get_thread_window(
-        self,
-        room_id: str,
-        thread_id: str,
-        *,
-        budget: ThreadReadBudget = UNBOUNDED_THREAD_READ,
-    ) -> ThreadWindowRead:
-        """Return cached events for one thread plus whether the window left anything out."""
         return await self._operation(
             room_id,
-            operation="get_thread_window",
-            disabled_result=ThreadWindowRead(events=None, truncated=False),
-            callback=lambda db: postgres_event_cache_threads.load_thread_window(
+            operation="get_thread_events",
+            disabled_result=None,
+            callback=lambda db: postgres_event_cache_threads.load_thread_events(
                 db,
                 namespace=self._runtime.namespace,
                 room_id=room_id,
                 thread_id=thread_id,
-                budget=budget,
             ),
         )
