+++ src/mindroom/matrix/cache/sqlite_event_cache_events.py
@@ -561,16 +561,14 @@ async def write_lookup_index_rows(
                 room_id,
                 origin_server_ts,
                 event_json,
-                event_bytes,
                 sender,
                 cached_at,
                 write_seq
             )
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)
+            VALUES (?, ?, ?, ?, ?, ?, ?, ?)
             ON CONFLICT(principal_id, room_id, event_id) DO UPDATE SET
                 origin_server_ts = excluded.origin_server_ts,
                 event_json = excluded.event_json,
-                event_bytes = excluded.event_bytes,
                 sender = excluded.sender,
                 cached_at = excluded.cached_at,
                 write_seq = excluded.write_seq
@@ -584,7 +582,6 @@ async def write_lookup_index_rows(
                 room_id,
                 event.origin_server_ts,
                 event.event_json,
-                event.event_bytes,
                 event.sender,
                 cached_at,
                 write_sequence,
