+++ tests/test_event_cache_storage_maintenance.py
@@ -585,15 +585,16 @@ async def test_postgres_reconnect_rejects_changed_certification_generation(
 
 
 @pytest.mark.asyncio
-async def test_bounded_read_columns_backfill_every_namespace_not_just_the_first(
+async def test_sender_backfills_in_every_namespace_not_just_the_first(
     postgres_event_cache_url: str,
 ) -> None:
-    """Each principal's legacy rows get priced, not only the one that wins the version bump.
+    """Each principal's legacy rows get a sender, not only the one that wins the version bump.
 
     ``schema_version`` is global to the database while every Matrix principal owns its own
     namespace. Gating the backfill on that shared version let the first principal to start
-    backfill itself, write the new version, and strand every other principal's rows at the column
-    defaults - which prices their events as free and makes ``max_bytes`` inert forever.
+    backfill itself, write the new version, and strand every other principal's rows at the ''
+    default - which makes every event look like the same author, so a collapsed read can no longer
+    tell an author's own edit from a foreign one and lets the foreign replacement win.
     """
     first = f"tenant_{uuid.uuid4().hex}"
     second = f"tenant_{uuid.uuid4().hex}"
@@ -605,9 +606,9 @@ async def test_bounded_read_columns_backfill_every_namespace_not_just_the_first(
                 await setup.execute(
                     """
                     INSERT INTO mindroom_event_cache_events(
-                        namespace, event_id, room_id, origin_server_ts, event_json, event_bytes, sender, cached_at
+                        namespace, event_id, room_id, origin_server_ts, event_json, sender, cached_at
                     )
-                    VALUES (%s, %s, %s, %s, %s, 0, '', 0)
+                    VALUES (%s, %s, %s, %s, %s, '', 0)
                     """,
                     (namespace, "$legacy", _ROOM_ID, 1000, '{"event_id":"$legacy","sender":"@a:localhost"}'),
                 )
@@ -623,13 +624,6 @@ async def test_bounded_read_columns_backfill_every_namespace_not_just_the_first(
         check = await psycopg.AsyncConnection.connect(database_url)
         try:
             for namespace in (first, second):
-                cursor = await check.execute(
-                    "SELECT event_bytes FROM mindroom_event_cache_events WHERE namespace = %s",
-                    (namespace,),
-                )
-                assert await cursor.fetchone() != (0,), f"{namespace} kept the event_bytes default"
-                await cursor.close()
-
                 cursor = await check.execute(
                     "SELECT sender FROM mindroom_event_cache_events WHERE namespace = %s",
                     (namespace,),
